+++ game/Wess/lcdload.cpp
@@ -188,9 +188,9 @@ int32_t wess_dig_lcd_data_read(
             // If uploading this sound would cause us to go beyond the bounds of SPU ram then do not try to upload this sound
             #if PSYDOOM_MODS
                 // PsyDoom: I think this condition was slightly wrong?
-                const bool bInsufficientSpuRam = ((int32_t) destSpuAddr + sndDataOffset + nextPatchSample.size > (int32_t) gPsxSpu_sram_end);
+                const bool bInsufficientSpuRam = (destSpuAddr + sndDataOffset + nextPatchSample.size > gPsxSpu_sram_end);
             #else
-                const bool bInsufficientSpuRam = ((int32_t) destSpuAddr + nextPatchSample.size > (int32_t)(gPsxSpu_sram_end + sndDataOffset));
+                const bool bInsufficientSpuRam = (destSpuAddr + nextPatchSample.size > gPsxSpu_sram_end + sndDataOffset);
             #endif
 
             if (bInsufficientSpuRam) {
